[CONTROLLER/TAGRECORDER] supports multi orgs in vtap-interfaces api
