+++ server/controller/http/server.go
@@ -109,7 +109,7 @@ func (s *Server) appendRegistrant() []registrant.Registrant {
 		router.NewVtapGroup(s.controllerConfig),
 		router.NewDataSource(s.controllerConfig),
 		router.NewVTapGroupConfig(s.controllerConfig),
-		router.NewVTapInterface(),
+		router.NewVTapInterface(s.controllerConfig.FPermit),
 		router.NewVtapRepo(),
 		router.NewPlugin(),
 		router.NewMail(),
