+++ server/controller/http/service/vtap_filter.go
@@ -17,80 +17,16 @@
 package service
 
 import (
-	"fmt"
-	"net/http"
-	"strconv"
-
 	"github.com/deepflowio/deepflow/server/controller/common"
 	"github.com/deepflowio/deepflow/server/controller/db/mysql"
-	"github.com/gin-gonic/gin"
+	httpcommon "github.com/deepflowio/deepflow/server/controller/http/common"
 )
 
-type UserInfo struct {
-	Type         int
-	ID           int
-	ORGID        int
-	DatabaseName string
-}
-
-func NewUserInfo(t, id, orgID int) *UserInfo {
-	return &UserInfo{
-		Type:  t,
-		ID:    id,
-		ORGID: orgID,
-	}
-}
-
-func GetUserInfo(c *gin.Context) *UserInfo {
-	orgID, _ := c.Get(common.HEADER_KEY_X_ORG_ID)
-	userType, _ := c.Get(common.HEADER_KEY_X_USER_TYPE)
-	userID, _ := c.Get(common.HEADER_KEY_X_USER_ID)
-	return &UserInfo{
-		Type:  userType.(int),
-		ID:    userID.(int),
-		ORGID: orgID.(int),
-	}
-}
-
-func GetUnauthorizedTeamIDs(userInfo *UserInfo, fpermitCfg *common.FPermit) (map[int]struct{}, error) {
-	if !fpermitCfg.Enabled {
-		return nil, nil
-	}
-
-	body := make(map[string]interface{})
-	response, err := common.CURLPerform(
-		http.MethodGet,
-		fmt.Sprintf(
-			"http://%s:%d/v1/org/%d/scope_teams?reverse=true",
-			fpermitCfg.Host, fpermitCfg.Port, userInfo.ORGID,
-		),
-		body,
-		common.WithHeader(common.HEADER_KEY_X_USER_TYPE, fmt.Sprintf("%d", userInfo.Type)),
-		common.WithHeader(common.HEADER_KEY_X_USER_ID, fmt.Sprintf("%d", userInfo.ID)),
-		common.WithHeader(common.HEADER_KEY_X_ORG_ID, fmt.Sprintf("%d", userInfo.ORGID)),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	teamIDMap := make(map[int]struct{}, 0)
-	for k, _ := range response.Get("DATA").MustMap() {
-		tid, err := strconv.Atoi(k)
-		if err != nil {
-			return nil, fmt.Errorf("%s: %s", k, err.Error())
-		}
-		teamIDMap[tid] = struct{}{}
-
-	}
-	log.Debugf("unauthorized team ids: %#v", teamIDMap)
-	return teamIDMap, nil
-}
-
-func getAgentByUser(userInfo *UserInfo, fpermitCfg *common.FPermit, vtaps []mysql.VTap) ([]mysql.VTap, error) {
+func getAgentByUser(userInfo *httpcommon.UserInfo, fpermitCfg *common.FPermit, vtaps []mysql.VTap) ([]mysql.VTap, error) {
 	if userInfo.Type == common.DEFAULT_USER_TYPE && userInfo.ID == common.DEFAULT_USER_ID {
 		return vtaps, nil
 	}
-	teamIDMap, err := GetUnauthorizedTeamIDs(userInfo, fpermitCfg)
+	teamIDMap, err := httpcommon.GetUnauthorizedTeamIDs(userInfo, fpermitCfg)
 	if err != nil {
 		return nil, err
 	}
@@ -111,11 +47,11 @@ func getAgentByUser(userInfo *UserInfo, fpermitCfg *common.FPermit, vtaps []mysq
 	return results, nil
 }
 
-func getAgentGroupByUser(userInfo *UserInfo, fpermitCfg *common.FPermit, vtapGroups []*mysql.VTapGroup) ([]*mysql.VTapGroup, error) {
+func getAgentGroupByUser(userInfo *httpcommon.UserInfo, fpermitCfg *common.FPermit, vtapGroups []*mysql.VTapGroup) ([]*mysql.VTapGroup, error) {
 	if userInfo.Type == common.DEFAULT_USER_TYPE && userInfo.ID == common.DEFAULT_USER_ID {
 		return vtapGroups, nil
 	}
-	teamIDMap, err := GetUnauthorizedTeamIDs(userInfo, fpermitCfg)
+	teamIDMap, err := httpcommon.GetUnauthorizedTeamIDs(userInfo, fpermitCfg)
 	if err != nil {
 		return nil, err
 	}
